+++ src/main/java/potato/potatoAPIserver/cart/service/CartProductService.java
@@ -1,5 +1,6 @@
 package potato.potatoAPIserver.cart.service;
 
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -8,53 +9,47 @@ import potato.potatoAPIserver.cart.domain.Cart;
 import potato.potatoAPIserver.cart.domain.CartProduct;
 import potato.potatoAPIserver.cart.dto.request.AddToCartRequest;
 import potato.potatoAPIserver.cart.repository.CartProductRepository;
-import potato.potatoAPIserver.cart.repository.CartRepository;
 import potato.potatoAPIserver.common.CustomException;
 import potato.potatoAPIserver.common.ResultCode;
 import potato.potatoAPIserver.order.repository.ProductRepository;
 import potato.potatoAPIserver.product.domain.Product;
-import potato.potatoAPIserver.user.domain.User;
-import potato.potatoAPIserver.user.repository.UserRepository;
 
 /**
  * @Author 허석문
- * @Since 2023-10-23
+ * @Since 2023-11-01
  */
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class CartWriteService {
+public class CartProductService {
 
-    private final UserRepository userRepository;
-    private final CartRepository cartRepository;
+    private final CartService cartService;
     private final CartProductRepository cartProductRepository;
     private final ProductRepository productRepository;
 
     public void addToCart(Long userId, AddToCartRequest request) {
-        Cart cart = cartRepository.findByUserId(userId)
-                .orElse(createCart(userId));
+        Cart cart = cartService.getCart(userId).orElse(
+                cartService.createCart(userId)
+        );
 
         Product product = productRepository.findById(request.getProductId())
                 .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, ResultCode.PRODUCT_NOT_FOUND));
 
-        CartProduct cartProduct = CartProduct
-                .builder()
-                .cart(cart)
-                .product(product)
-                .quantity(request.getQuantity())
-                .build();
+        CartProduct cartProduct = cartProductRepository.findCart(cart.getId(), product.getId()).orElse(null);
+
+        if (cartProduct != null) {
+            // 상품이 이미 장바구니에 존재하는 경우
+            cartProduct.addQuantity(request.getQuantity());
+        } else {
+            // 상품이 장바구니에 없는 경우
+            cartProduct = CartProduct
+                    .builder()
+                    .cart(cart)
+                    .product(product)
+                    .quantity(request.getQuantity())
+                    .build();
+        }
 
         cartProductRepository.save(cartProduct);
-    }
-
-    public Cart createCart(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, ResultCode.USER_NOT_FOUND));
-
-        Cart cart = Cart.builder()
-                .user(user)
-                .build();
-
-        return cartRepository.save(cart);
     }
 }
