+++ src/main/java/potato/potatoAPIserver/cart/service/CartService.java
@@ -1,0 +1,42 @@
+package potato.potatoAPIserver.cart.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import potato.potatoAPIserver.cart.domain.Cart;
+import potato.potatoAPIserver.cart.repository.CartRepository;
+import potato.potatoAPIserver.common.CustomException;
+import potato.potatoAPIserver.common.ResultCode;
+import potato.potatoAPIserver.user.domain.User;
+import potato.potatoAPIserver.user.repository.UserRepository;
+
+import java.util.Optional;
+
+/**
+ * @Author 허석문
+ * @Since 2023-10-23
+ */
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class CartService {
+
+    private final CartRepository cartRepository;
+    private final UserRepository userRepository;
+
+    public Optional<Cart> getCart(Long userId) {
+        return cartRepository.findByUserId(userId);
+    }
+
+    public Cart createCart(Long userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new CustomException(HttpStatus.NOT_FOUND, ResultCode.USER_NOT_FOUND));
+
+        Cart cart = Cart.builder()
+                .user(user)
+                .build();
+
+        return cartRepository.save(cart);
+    }
+}
